ДЗ 5: сделал разбивку скрипта по модулям
