+++ js/redactor.js
@@ -1,0 +1,1 @@
+// файл под работу с фото
