+++ js/upload.js
@@ -1,0 +1,1 @@
+// файл под логику загрузки фото
